Rename package for rxjava2 compat
