+++ build.gradle
@@ -29,7 +29,7 @@ ext {
     bintrayRepo = 'tbruyelle'
     bintrayName = 'RxPermissions'
 
-    publishedGroupId = 'com.tbruyelle.rxpermissions'
+    publishedGroupId = 'com.tbruyelle.rxpermissions2'
     artifact = 'rxpermissions'
     libraryName = 'RxPermissions'
     libraryVersion = '0.7.0'
