+++ lib/src/main/java/com/tbruyelle/rxpermissions2/Permission.java
@@ -1,4 +1,4 @@
-package com.tbruyelle.rxpermissions;
+package com.tbruyelle.rxpermissions2;
 
 public class Permission {
     public final String name;
