+++ lib/src/test/java/com/tbruyelle/rxpermissions2/RxPermissionsTest.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.tbruyelle.rxpermissions;
+package com.tbruyelle.rxpermissions2;
 
 import android.Manifest;
 import android.annotation.TargetApi;
@@ -25,9 +25,7 @@ import io.reactivex.Observable;
 import io.reactivex.observers.TestObserver;
 import io.reactivex.subjects.PublishSubject;
 import org.junit.*;
-import org.junit.runner.RunWith;
 import org.mockito.*;
-import org.mockito.runners.MockitoJUnitRunner;
 
 import static org.junit.Assert.*;
 import static org.mockito.Matchers.anyString;
