+++ sample/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
-    package="com.tbruyelle.rxpermissions.sample"
+    package="com.tbruyelle.rxpermissions2.sample"
     xmlns:android="http://schemas.android.com/apk/res/android"
     >
 
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.AppCompat.Light.DarkActionBar"
         >
         <activity
-            android:name="com.tbruyelle.rxpermissions.sample.MainActivity"
+            android:name="com.tbruyelle.rxpermissions2.sample.MainActivity"
             android:label="@string/app_name"
             >
             <intent-filter>
