+++ settings.gradle
@@ -1,3 +1,3 @@
 include 'lib'
-project(':lib').name = 'rxpermissions'
+project(':lib').name = 'rxpermissions2'
 include 'sample'
